requirements.txt
alembic==1.10.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.0.8
async-generator==1.10
attrs==21.4.0
autobahn==22.7.1
Automat==22.10.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==23.1.0
bleach==5.0.0
boto3==1.26.3
botocore==1.29.3
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
cleo==2.0.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
crashtest==0.4.1
cryptography==39.0.1
daphne==4.0.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==3.2.13
django-allauth==0.52.0
django-appconf==1.0.5
django-bootstrap5==22.1
django-cors-headers==3.14.0
django-environ==0.9.0
django-extensions==3.2.1
django-imagekit==4.1.0
django-phonenumber-field==7.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-storages==1.13.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
dulwich==0.20.50
entrypoints==0.4
exceptiongroup==1.0.2
executing==1.1.1
fastapi==0.95.0
fastjsonschema==2.15.3
filelock==3.9.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httptools==0.5.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.0.0
incremental==22.10.0
ipykernel==6.14.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jaraco.classes==3.2.3
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keyring==23.13.1
lockfile==0.12.2
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==0.8.4
more-itertools==9.0.0
msgpack==1.0.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
oauthlib==3.2.2
outcome==1.2.0
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
phonenumbers==8.12.57
phonenumberslite==8.12.57
pi==0.1.2
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
pkginfo==1.9.6
platformdirs==3.0.0
poetry==1.3.2
poetry-core==1.4.0
poetry-plugin-export==1.3.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.1
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.6
Pygments==2.13.0
PyJWT==1.7.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7.1
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.5
PyYAML==6.0
pyzmq==23.1.0
qtconsole==5.3.1
QtPy==2.1.0
rapidfuzz==2.13.7
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
s3transfer==0.6.0
selenium==4.6.0
Send2Trash==1.8.0
service-identity==21.1.0
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.7
sqlparse==0.4.3
stack-data==0.5.1
starlette==0.26.1
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.1
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
trove-classifiers==2023.1.20
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
uvicorn==0.21.1
virtualenv==20.17.1
watchfiles==0.18.1
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.4
whitenoise==6.2.0
widgetsnbextension==3.6.0
wsproto==1.2.0
zipp==3.12.0
zope.interface==5.5.2
